backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
langchain
langchain-community
langchain-openai
redis
vllm
transformers
accelerate
sentencepiece
protobuf